requirements.txt
Django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
django-storages[boto3]>=1.14
django-cors-headers>=4.3
python-dotenv>=1.0
Pillow>=10.0
drf-spectacular>=0.27
boto3>=1.34
celery>=5.3
redis>=5.0
django-redis>=5.4
minio>=7.2
PyJWT>=2.8
